centralize route definitions
